+++ src/main/java/uz/pdp/notuzum/DTO/LoginRequest.java
@@ -1,0 +1,18 @@
+package uz.pdp.notuzum.DTO;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginRequest {
+    private String username;
+    private String password;
+
+
+
+}